+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+ - Input Format option for TextField and TextArea components.
+
 ## [Unreleased]
 ### Added
  - Multiple masks settings for text field and phone number components