+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.35.6
+### Fixed
+ - The PDF builder to not delete any changes to the form that are made.
+
 ## 2.35.4, 2.35.5
 ### Fixed
  - The PDF builder to work with cross-site forms.